Extract SignedDocumentEditor, fix default validity period
