+++ cli/src/ui/component.rs
@@ -13,6 +13,7 @@ pub enum ComponentStatus {
     Escaped,
 }
 
+#[derive(Debug)]
 pub struct CursorPosition {
     pub x: u16,
     pub y: u16,
