+++ src/validator.rs
@@ -15,7 +15,7 @@ use crate::{
         validity_period::validator::{validate_timestamp, validate_validity_period},
         SIGNED_CERTIFICATE_SCHEMA_ID, SIGNED_NODE_DESCRIPTOR_SCHEMA_ID,
     },
-    Error,
+    Error, Result,
 };
 
 use self::validated_data::{ValidatedCertificate, ValidatedNodeDescriptor};
@@ -29,7 +29,7 @@ pub mod validated_data;
 pub fn validate_certificate_str(
     data: &str,
     timestamp: Option<DateTime<Utc>>,
-) -> Result<ValidatedCertificate, Error> {
+) -> Result<ValidatedCertificate> {
     let value: Value = serde_json::from_str(data).map_err(|e| Error::InvalidJson(e.to_string()))?;
     validate_certificate(value, timestamp)
 }
@@ -41,7 +41,7 @@ pub fn validate_certificate_str(
 pub fn validate_certificate(
     value: Value,
     timestamp: Option<DateTime<Utc>>,
-) -> Result<ValidatedCertificate, Error> {
+) -> Result<ValidatedCertificate> {
     validate_schema(&value, SIGNED_CERTIFICATE_SCHEMA_ID, "certificate")?;
     let signed_certificate: SignedCertificate = serde_json::from_value(value)
         .map_err(|e| Error::JsonDoesNotConformToSchema(e.to_string()))?;
@@ -52,12 +52,12 @@ pub fn validate_certificate(
     Ok(validated_certificate)
 }
 
-pub fn validate_node_descriptor_str(data: &str) -> Result<ValidatedNodeDescriptor, Error> {
+pub fn validate_node_descriptor_str(data: &str) -> Result<ValidatedNodeDescriptor> {
     let value: Value = serde_json::from_str(data).map_err(|e| Error::InvalidJson(e.to_string()))?;
     validate_node_descriptor(value)
 }
 
-pub fn validate_node_descriptor(value: Value) -> Result<ValidatedNodeDescriptor, Error> {
+pub fn validate_node_descriptor(value: Value) -> Result<ValidatedNodeDescriptor> {
     validate_schema(&value, SIGNED_NODE_DESCRIPTOR_SCHEMA_ID, "node descriptor")?;
     let signed_node_descriptor: SignedNodeDescriptor = serde_json::from_value(value)
         .map_err(|e| Error::JsonDoesNotConformToSchema(e.to_string()))?;
@@ -68,7 +68,7 @@ pub fn validate_node_descriptor(value: Value) -> Result<ValidatedNodeDescriptor,
     Ok(validated_node_descriptor)
 }
 
-fn validate_schema(value: &Value, schema_id: &str, structure_name: &str) -> Result<(), Error> {
+fn validate_schema(value: &Value, schema_id: &str, structure_name: &str) -> Result<()> {
     value["$schema"]
         .as_str()
         .map(|schema| {
@@ -90,7 +90,7 @@ fn validate_schema(value: &Value, schema_id: &str, structure_name: &str) -> Resu
 
 fn validate_signed_node_descriptor(
     signed_node_descriptor: SignedNodeDescriptor,
-) -> Result<ValidatedNodeDescriptor, Error> {
+) -> Result<ValidatedNodeDescriptor> {
     let node_descriptor: NodeDescriptor =
         serde_json::from_value(signed_node_descriptor.node_descriptor.clone())
             .map_err(|e| Error::JsonDoesNotConformToSchema(e.to_string()))?;
@@ -128,11 +128,11 @@ fn validate_signed_node_descriptor(
 
 fn create_certificate_fingerprint(
     signed_certificate: &SignedCertificate,
-) -> Result<Fingerprint, Error> {
+) -> Result<Fingerprint> {
     create_fingerprint_for_value(&signed_certificate.certificate)
 }
 
-fn create_fingerprint_for_value(value: &Value) -> Result<Fingerprint, Error> {
+fn create_fingerprint_for_value(value: &Value) -> Result<Fingerprint> {
     create_default_hash(value).map(|binary| binary.encode_hex())
 }
 
@@ -144,7 +144,7 @@ fn create_fingerprint_for_value(value: &Value) -> Result<Fingerprint, Error> {
 fn validate_signed_certificate(
     signed_certificate: &SignedCertificate,
     timestamp: Option<DateTime<Utc>>,
-) -> Result<ValidatedCertificate, Error> {
+) -> Result<ValidatedCertificate> {
     let parent = match &signed_certificate.signature.signer {
         Signer::SelfSigned => {
             let certificate: Certificate =
